ThemeSong class was created to restructure the code properly
